backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic-settings==2.8.1
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.1
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
httpx==0.28.1
pytest==8.3.5
